# Conflicts:
#	yeshtery/src/main/java/com/nasnav/yeshtery/controller/v1/YeshteryController.java
#	yeshtery/src/test/java/com/nasnav/yeshtery/test/YeshteryApiTest.java

+++ nasnav-services/src/main/java/com/nasnav/dto/CategoryDTO.java
@@ -22,7 +22,7 @@ public class CategoryDTO {
 
         @Schema(name = "Category's parent ID", example = "123", required = false)
         @JsonProperty("parent_id")
-        public Integer parentId;
+        public Long parentId;
 
         @Schema(name = "Category's logo", example = "/categories/logos/564961451_56541.jpg", required = false)
         @JsonProperty("logo")
